discover/devmapper: Use explicit 64-bit types for sector counts
We currently have a bug where the return value from get_block_sectors
may overflow an unsigned int, so we create a snapshot that is too small.

This change uses uint64_t types for the sector counts.

+++ discover/device-handler.h
@@ -65,7 +65,7 @@ struct ramdisk_device {
 	char		*snapshot;
 	char		*origin;
 	char		*base;
-	unsigned int	sectors;
+	uint64_t	sectors;
 };
 
 struct device_handler *device_handler_init(struct discover_server *server,
